requirements.txt
django
django-haystack